+++ src/Server/Helper.php
@@ -167,10 +167,6 @@ class Helper
             $errors[] = new RequestError('GraphQL Request must include at least one of those two parameters: "query" or "queryId"');
         }
 
-        if ($query !== '' && $queryId !== '') {
-            $errors[] = new RequestError('GraphQL Request parameters "query" and "queryId" are mutually exclusive');
-        }
-
         if (! is_string($query)) {
             $errors[] = new RequestError(
                 'GraphQL Request parameter "query" must be string, but got ' .
@@ -279,9 +275,9 @@ class Helper
                 );
             }
 
-            $doc = ($op->queryId ?? '') === ''
-                ? $op->query
-                : $this->loadPersistedQuery($config, $op);
+            $doc = $op->queryId !== null && $op->query === null
+                ? $this->loadPersistedQuery($config, $op)
+                : $op->query;
 
             if (! $doc instanceof DocumentNode) {
                 $doc = Parser::parse($doc);